adding polynomial solver
