+++ kornia/geometry/solvers/__init__.py
@@ -1,0 +1,9 @@
+from .polynomial_solver import (
+    solve_quadratic,
+    solve_cubic,
+)
+
+__all__ = [
+    "solve_quadratic",
+    "solve_cubic",
+]
